Allow to collect debug for all DIMMs in the system.
This change allows the user to collect debug logs from all the DIMMs in
a system in one command. Currently the user needs to get all DIMMs IDs
using "show -dimm" and then iterate though those calling "dump -debug".
While this is doable it requires complicated shell scripts.

With this change the user passes a folder as destination. When -dimm
argument is empty, the tool will select all DIMMs in the system. It is
still possible to select a subset of DIMMs. For each DIMM, a file with
the serial number is placed on the provided folder.

Signed-off-by: Luis Alberto Herrera <luisalberto@google.com>